test: use deterministic guard resolution in mini-router test
