+++ packages/nuxt/src/app/plugins/router.ts
@@ -128,14 +128,12 @@ const plugin: Plugin<{ route: Route, router: Router }> & ObjectPlugin<{ route: R
     const route: Route = reactive(getRouteFromPath(initialURL))
     let navigationCounter = 0
     async function handleNavigation (url: string | Partial<Route>, replace?: boolean): Promise<void> {
-      // Track this navigation so a later one can supersede it (last-navigation-wins).
       const navigationId = ++navigationCounter
       try {
         // Resolve route
         const to = getRouteFromPath(url)
 
-        // Run beforeEach hooks. Bail as soon as a later navigation supersedes
-        // this one so superseded navigations stop running middleware (#31762).
+        // Run beforeEach hooks, bailing if a later navigation supersedes this one (#31762)
         for (const middleware of hooks['navigate:before']) {
           const result = await middleware(to, route)
           if (navigationId !== navigationCounter) { return }
